+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Type Svg renamed to SVG
+- Type SvgWriter renamed to SVGWriter
 
 ## [0.11.0] 2020-03-17
 ### Changed